Fixed dataset importer

+++ backend/src/labelit/services/dataset_importer.py
@@ -8,10 +8,10 @@ from labelit.models import Document, Dataset
 
 class DatasetImporter:
     def __init__(
-        self,
-        path_to_uploaded_directory: str,
+        self, path_to_uploaded_directory: str, allow_not_present_files: bool = True
     ):
         self.dir_path = path_to_uploaded_directory
+        self.allow_not_present_files = allow_not_present_files
 
     def _validate(self):
         # TODO: implement
@@ -44,17 +44,26 @@ class DatasetImporter:
                     dataset=dataset,
                 )
                 audio_name = f"{doc_name}.wav"
+                metadata_name = f"{doc_name}.meta.json"
                 text_name = f"{doc_name}.txt"
-                if audio_name in filenames:
+
+                with open(os.path.join(documents_dir_path, metadata_name), "r") as f:
+                    metadata = json.load(f)
+
+                if audio_name in filenames or self.allow_not_present_files:
                     doc_dict["audio_filename"] = audio_name
                     source_audio_path = os.path.join(documents_dir_path, audio_name)
-                    doc_dict["audio_duration"] = 1000 * get_audio_duration_in_seconds(
-                        source_audio_path
-                    )
+                    if metadata.get("audio_duration", None):
+                        doc_dict["audio_duration"] = metadata["audio_duration"]
+                    else:
+                        doc_dict[
+                            "audio_duration"
+                        ] = 1000 * get_audio_duration_in_seconds(source_audio_path)
 
                 if text_name in filenames:
                     with open(os.path.join(documents_dir_path, text_name)) as f:
                         doc_dict["text"] = f.read()
+
                 Document.objects.create(**doc_dict)
 
         print("Importing documents...")
@@ -66,15 +75,16 @@ class DatasetImporter:
             audio_doc_names = list(
                 dataset_docs.values_list("audio_filename", flat=True)
             )
+
             for idx, audio_name in enumerate(audio_doc_names):
                 print(f"Audio import: {idx}/{len(audio_doc_names)}")
                 audio_path = os.path.join(self.dir_path, "documents", audio_name)
-                # logger.debug(f"&type of audio_storage.open(audio_path, 'wb').obj: {type(audio_storage.open(audio_path, 'wb').obj)}")
-                # audio_storage.open(audio_path, "rb").obj.upload_file(f"{audio_name}/")
-                with open(audio_path, "rb") as file_reader:
-                    writer = audio_storage.open(audio_name, "w")
-                    writer.write(file_reader.read())
-                    writer.close()
+
+                if os.path.exists(audio_path):
+                    with open(audio_path, "rb") as file_reader:
+                        writer = audio_storage.open(audio_name, "w")
+                        writer.write(file_reader.read())
+                        writer.close()
 
         print("Uploading audios...")
         _upload_audio_to_storage()
